factor out stream code from tcp
so it can be used by unix sockets
