+++ glommio/src/net/mod.rs
@@ -77,6 +77,7 @@ fn yolo_sendmsg(
     }
 }
 
+mod stream;
 mod tcp_socket;
 mod udp_socket;
 pub use self::tcp_socket::{AcceptedTcpStream, TcpListener, TcpStream};
